move test implementations inside of StructureTest struct. file renaming/refactoring for clarity
