+++ check_if_image_tag_exists/cloudbuild.yaml.in
@@ -2,7 +2,7 @@ steps:
   - name: gcr.io/cloud-builders/docker
     args: ['build', '--tag=${IMAGE}', '.']
     id: BUILD
-  - name: gcr.io/gcp-runtimes/structure_test
+  - name: gcr.io/nick-cloudbuild/structure_test:refactor
     args: ['--image', '${IMAGE}', '-v', '--config', '/workspace/test_config.json']
     id: STRUCTURE_TEST
 images: ['${IMAGE}']
