+++ structure_tests/cloudbuild.yaml.in
@@ -3,6 +3,6 @@ steps:
           args: ['test', '-c', 'github.com/GoogleCloudPlatform/runtimes-common/structure_tests', '-o', '/workspace/structure_tests/structure_test']
           env: ['PROJECT_ROOT=github.com/GoogleCloudPlatform/runtimes-common']
         - name: gcr.io/cloud-builders/docker
-          args: ['build', '-t', 'gcr.io/gcp-runtimes/structure_test:${VERSION}', './structure_tests']
+          args: ['build', '-t', 'gcr.io/nick-cloudbuild/structure_test:${VERSION}', './structure_tests']
 images:
-        - 'gcr.io/gcp-runtimes/structure_test:${VERSION}'
+        - 'gcr.io/nick-cloudbuild/structure_test:${VERSION}'
