+++ structure_tests/file_content_test_v0.go
@@ -1,12 +1,5 @@
 package main
 
-type FileContentTest interface {
-	GetName() string               // name of test
-	GetPath() string               // file to check existence of
-	GetExpectedContents() []string // list of expected contents of file
-	GetExcludedContents() []string // list of excluded contents of file
-}
-
 type FileContentTestv0 struct {
 	Name             string   // name of test
 	Path             string   // file to check existence of
